+++ integrations/jira/CHANGELOG.md
@@ -7,19 +7,36 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- towncrier release notes start -->
 
-0.1.2 (2023-08-11)
+0.1.5 (2023-09-27)
+
+### Improvements
+
+- Bumped ocean to version 0.3.1 (#1)
+
+  0.1.4 (2023-09-13)
+
+### Improvements
+
+- Bumped ocean to 0.3.0 (#1)
+
+  0.1.3 (2023-08-29)
+
+### Improvements
+
+- Changed the app_host to not be required for the installation (PORT-4527)
+- Bumped Ocean to 0.2.3 (#1)
+
+  0.1.2 (2023-08-11)
 
 ### Improvements
 
 - Optimized dockerfile to produce smaller images (PORT-4485)
 
-
-0.1.1 (2023-08-11)
+  0.1.1 (2023-08-11)
 
 ### Improvements
 
 - Upgraded ocean to version 0.2.2
-
 
 v0.1.0 (2023-08-10)