+++ integrations/newrelic/CHANGELOG.md
@@ -7,24 +7,38 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- towncrier release notes start -->
 
-# Port_Ocean 0.1.2 (2023-08-11)
+#  0.1.5 (2023-10-05)
+
+## Bug Fixes
+
+- Fixed unhandled exception when no entity is returned for a given entity guid found in an newRelic Alert (#1)
+
+#  0.1.4 (2023-09-27)
+
+### Improvements
+
+- Bumped ocean to version 0.3.1 (#1)
+
+#  0.1.3 (2023-09-13)
+
+### Features
+
+- Changed returned entities to ocean with a generator of a list to support ocean 0.3.0
+
+# 0.1.2 (2023-08-11)
 
 ### Improvements
 
 - Optimized dockerfile to produce smaller images (PORT-4485)
 
-
-0.1.1 (2023-08-03)
-==================
+#  0.1.1 (2023-08-03)
 
 ### Bug Fixes
 
 - Add README.md (#1)
 - Support specifying whether to count open issues per new relic entity (#1)
 
-
-0.1.0 (2023-08-02)
-==================
+#  0.1.0 (2023-08-02)
 
 ### Features